Separate Device&Context from selector, apply other comments
